wip(atomic) : it compile ; doesn't work for now, need to fix pipe refresh / sync

+++ Holovibes/includes/API/contrast_API.hh
@@ -22,7 +22,7 @@ void set_auto_contrast_all();
 
 bool set_auto_contrast();
 
-void set_current_window_contrast_min(const double value);
-void set_current_window_contrast_max(const double value);
+void set_current_window_contrast_min(const float value);
+void set_current_window_contrast_max(const float value);
 
 } // namespace holovibes::api
